fix(encoder): address PR review feedback on mode-dependent defaults and velocity window
- config.h: ENCODER_TICKS_PER_REV now conditional on ENCODER_USE_FULL_STEP
  (20 in full-step mode, 80 in raw x4 mode) so DEFAULT_CAL_FACTOR is
  always correct regardless of which mode is compiled in
- config.h: DEFAULT_MIN_EXT_VEL 0.5f → 0.3f (more appropriate for slow printing)
- config.h: VEL_MEDIAN_N 15 → 525 — window now captures ≥ 2 movement ticks
  at DEFAULT_MIN_EXT_VEL (0.3 mm/s); derivation documented in comment
- encoder.cpp: s_win_idx uint8_t → uint16_t (VEL_MEDIAN_N=525 exceeds 255)
- encoder.cpp: accumulation loop index widened to uint16_t to match

Agent-Logs-Url: https://github.com/tobi01001/klipper-filament-runout-esp32/sessions/b70c24be-0ffe-44b8-b099-af954f240162

+++ firmware/src/encoder.cpp
@@ -176,9 +176,9 @@ void encoder_task(void * /*param*/) {
         // spike across N windows suppresses it adequately.
         const float cal_factor = s_cfg ? s_cfg->cal_factor : DEFAULT_CAL_FACTOR;
 
-        static int32_t  s_win_ticks[VEL_MEDIAN_N] = {};
-        static uint32_t s_win_dt_ms[VEL_MEDIAN_N]  = {};
-        static uint8_t  s_win_idx = 0;
+        static int32_t   s_win_ticks[VEL_MEDIAN_N] = {};
+        static uint32_t  s_win_dt_ms[VEL_MEDIAN_N]  = {};
+        static uint16_t  s_win_idx = 0;  // uint16_t: VEL_MEDIAN_N can exceed 255
 
         s_win_ticks[s_win_idx] = delta;
         s_win_dt_ms[s_win_idx] = dt_ms;
@@ -186,7 +186,7 @@ void encoder_task(void * /*param*/) {
 
         int32_t  win_tick_sum = 0;
         uint32_t win_time_ms  = 0;
-        for (uint8_t i = 0; i < VEL_MEDIAN_N; ++i) {
+        for (uint16_t i = 0; i < VEL_MEDIAN_N; ++i) {
             win_tick_sum += s_win_ticks[i];
             win_time_ms  += s_win_dt_ms[i];
         }
